refactor: move sending current rate logic from task schedule service to email service
