+++ src/task-schedule/task-schedule.constants.ts
@@ -1,3 +1,1 @@
-export const EMAIL_SUBJECT = 'Daily Exchange Rate Report: USD to UAH';
-
 export const TIMEZONE = 'Europe/Kyiv';
